Fix store naming
